picking: a pick id is an identity — three carriers said otherwise
The 2^24 wall was still standing in the subsystem that resolves every hover and click,
and raising the arena ceiling walked it into range rather than leaving it behind.

A pick id must survive four carriers unchanged: the JS allocator, the shader uniform,
the shader's RGBA8 byte packing, and the CPU decode. Three of the four were wrong.

  1. baseId was `uniform(0)` — a JS number, which is an f32 uniform. Exact only to
     2^24, so base 16,777,217 arrives as 16,777,216 and two glyphs answer to one id.
     Measured: ARENA_MAX_BYTES itself aliases, 44,739,242 -> 44,739,240.
     The comment above it read "int-cast the (float) per-object uniform so the bit ops
     stay exact" — the int cast makes the bit ops exact and cannot recover precision
     the f32 already discarded. The defect was documented and misread as handled.

  2. `int(baseId).add(int(instanceIndex))` narrowed to signed i32. Two costs, not one:
     it caps the space at 2^31, AND it makes shiftRight ARITHMETIC rather than logical,
     so the sign bit smears through the alpha byte for any id >= 2^31.

  3. instancePickingId was a Float32Array in both GlyphField and MegaGlyphField. No
     shader reads it — it is the CPU-side mirror harnesses check the pick pass against.
     As an f32 it aliased past 2^24 while the shader (once fixed) did not, so the
     mirror would have disagreed with the thing it exists to verify.

The decode was already correct: JS numbers are f64, and a*2^24 + (r<<16|g<<8|b) stays
exact across the whole u32 range. Nothing to fix, verified rather than assumed.

THE GUARD WAS WRONG IN BOTH DIRECTIONS

It warned above 0x7FFFFFFF: 128x too permissive against the f32 uniform's true 2^24
limit, and a WARN for a condition whose entire symptom is that the app silently acts on
the wrong target. It refuses now, at the true u32 bound, naming the block and the live
block count.

WHY THIS WAS REACHABLE

Pick ids are bounded by live glyph count. At the old 16.7MB arena they sat right at
2^24 and essentially never crossed. At ARENA_MAX_BYTES = 44,739,242 they cross it by
2.7x. Third instance of this pattern in a week (slot ordinals, foldScalars totalRows,
pick ids) — the ceiling raise did not leave these behind, it walked each of them into a
wall it had been too small to hit.

tools/pick-identity.test.mjs — 35 teeth. The pack/unpack round trip across the u32
space including both boundaries; a positive assertion that f32 genuinely IS lossy here
(if it weren't, this migration was unnecessary and the file is noise); the source-level
carrier checks; and the guard's exactness and refusal.

Mutation-tested five ways, all caught: float uniform, int() narrowing, guard back to
warn-at-31-bits, and the CPU mirror at each of its two declaration sites.

That last one is why the count is five and not four. The first pass used a single-line
regex, and GlyphField's PRIMARY declaration wraps across two lines — mutating that site
alone left the file reporting 33/33 with the defect fully reinstated, because the resize
path still matched and the "declares it" tooth only checked for nonzero matches. The
regex spans lines now and the site count is pinned, so a declaration added out of reach
lowers the count instead of quietly lowering coverage. A grep-based guard's blind spot
is invisible from inside the guard; only the mutation showed it.

A related trap in the same file, kept as a comment: the guard's "no 31-bit bound
survives" tooth first failed on the commit's own explanatory comment, which cites
0x7FFFFFFF to say why it was wrong. It matches the comparison now, not the constant —
a source check that cannot tell code from prose fails on its documentation, or passes
on a stale one.

STILL OWED: no GPU gate renders a pick pass at an id past 2^24 and asserts the readback
resolves to the glyph it aimed at. The shader carriers are checked at the source level
only. arrows-check (45 assertions, real GPU) exercises picking and passes, but not in
that range. The file says so rather than implying coverage it lacks.

Verified: 50/50 test files, all six GPU gates 0 failures on apple/metal-3 with vite
confirmed serving the new source, app build clean.

+++ packages/glyph3d-core/src/GlyphField.js
@@ -1140,9 +1140,12 @@ export default class GlyphField {
         }
         geometry.setAttribute('instanceGroupId',
             new THREE.InstancedBufferAttribute(new Float32Array(maxCount), 1));
-        // instancePickingId — written by PickingSystem.register() after flush
+        // instancePickingId — written by PickingSystem.register() after flush.
+        // Uint32Array: a pick id is an exact identity and f32 aliases it past 2^24.
+        // No shader reads this attribute; it is the CPU-side mirror harnesses check
+        // the pick pass against, so it must carry ids exactly or it lies about them.
         geometry.setAttribute('instancePickingId',
-            new THREE.InstancedBufferAttribute(new Float32Array(maxCount), 1));
+            new THREE.InstancedBufferAttribute(new Uint32Array(maxCount), 1));
 
         geometry._maxInstanceCount = maxCount;
         geometry.instanceCount = 0;
@@ -2197,7 +2200,7 @@ export default class GlyphField {
         geom.setAttribute('instanceColor',    new THREE.InstancedBufferAttribute(col4, 4, true));
         geom.setAttribute('instanceGroupId',  new THREE.InstancedBufferAttribute(groupIds || new Float32Array(count), 1));
         if (!geom.attributes.instancePickingId || geom.attributes.instancePickingId.array.length < count) {
-            geom.setAttribute('instancePickingId', new THREE.InstancedBufferAttribute(new Float32Array(count), 1));
+            geom.setAttribute('instancePickingId', new THREE.InstancedBufferAttribute(new Uint32Array(count), 1));
         }
 
         this._ensureHighlightTexture(count);
